Change the rendering logic of the elements
The logic of rendering elements on the page
has been adjusted:
---

* added a condition for determining the presence
or absence of movies, based on which the desired
component is drawn.

* added an additional condition for drawing the
"Show more" button;

* changed the calculation of the number of movie
cards remaining to be shown.

A small refactoring of the code was also made:
the "magic" values were removed, some constants
were formed in the Glossaries.